build: remove nightly channel specific date
The date specification on the nightly channel was necessary because the unstable const_trait feature requires the effects feature gate, which in turn implies the compiler flag -Znext-solver. However, due to its instability and the removal of ~const and impl const from the standard library for a more refined implementation, the next-solver flag also causes compatibility issues with certain dependencies. To continue utilizing the latest Rust features, we decided to remove all const_trait and some const methods to maintain compatibility with the most recent toolchain.